move try_unwrap.rs and print.rs into utils.rs
