+++ src/utils.rs
@@ -1,9 +1,44 @@
+//! General purpose utilities.
+
 use colored::*;
 
 use std::cmp::min;
 use std::fmt;
 use std::iter;
 use std::time::Duration;
+
+// ---
+// --- TryUnwrap: wrapper for Option and Result
+// ---
+
+pub trait TryUnwrap {
+    type Val;
+    fn try_unwrap(self) -> Result<Self::Val, String>;
+}
+
+impl<T> TryUnwrap for Option<T> {
+    type Val = T;
+
+    fn try_unwrap(self) -> Result<Self::Val, String> {
+        if let Some(val) = self {
+            Ok(val)
+        } else {
+            Err("empty output".to_string())
+        }
+    }
+}
+
+impl<T, E: fmt::Display> TryUnwrap for Result<T, E> {
+    type Val = T;
+
+    fn try_unwrap(self) -> Result<Self::Val, String> {
+        self.map_err(|err| format!("{}", err))
+    }
+}
+
+// ---
+// --- Line: helper struct for printing
+// ---
 
 const DEFAULT_WIDTH: usize = 30;
